изменения: - панель Bank Mini Panel меняет ширину в зависимости от суммы денег - деньги теперь long вместо int - проверка подключения к серверу каждые 5с

+++ Assets/AssetStore/FortuneWheels/Scripts/Sector.cs
@@ -86,7 +86,7 @@ namespace MkeyFW
 
         private void MoneyToBank()
         {
-            var income = new Dictionary<MoneyType, int>();
+            var income = new Dictionary<MoneyType, long>();
             income.Add(moneyType, coins);
             MoneyManager.Instance.PlusMoney(income);
         }
